Keep shortcut unbinding inside Swift file budgets
The CI guard failed because the regression tests and recorder UI grew existing budgeted Swift files. Move the new shortcut tests into a focused test file and extract the recorder UI into its own source file so the existing budgeted files shrink while keeping the same behavior covered.

Constraint: Do not skip or weaken CI checks; fix the code shape that violated the guard.

Constraint: Local tests are not run per repository policy.

Rejected: Refresh the Swift file length budget | this would accept avoidable file growth instead of reducing it.

Confidence: high

Scope-risk: narrow

Directive: Keep future shortcut recorder UI work in Sources/KeyboardShortcutRecorder.swift instead of growing KeyboardShortcutSettings.swift again.

Tested: python3 scripts/swift_file_length_budget.py --budget .github/swift-file-length-budget.tsv; git diff --cached --check; plutil -lint GhosttyTabs.xcodeproj/project.pbxproj

Not-tested: Local unit/UI tests per repository policy; GitHub Actions will run the test suites.
