+++ Sources/KeyboardShortcutSettingsFileStore.swift
@@ -833,16 +833,12 @@ final class CmuxSettingsFileStore {
         _ rawValue: Any,
         action: KeyboardShortcutSettings.Action
     ) -> StoredShortcut? {
-        let shortcut: StoredShortcut?
-        if rawValue is NSNull {
-            shortcut = .unbound
-        } else if let stroke = jsonString(rawValue) {
-            shortcut = StoredShortcut.parseConfig(stroke)
-        } else if let strokes = jsonStringArray(rawValue) {
-            shortcut = StoredShortcut.parseConfig(strokes: strokes)
-        } else {
-            shortcut = nil
-        }
+        let shortcut: StoredShortcut? = {
+            if rawValue is NSNull { return .unbound }
+            if let stroke = jsonString(rawValue) { return StoredShortcut.parseConfig(stroke) }
+            if let strokes = jsonStringArray(rawValue) { return StoredShortcut.parseConfig(strokes: strokes) }
+            return nil
+        }()
 
         guard let shortcut else { return nil }
         if let normalized = action.normalizedRecordedShortcut(shortcut) {
